Refactor Subsocial API connection
