+++ src/postgres/event-handlers/AccountFollowed.ts
@@ -1,8 +1,8 @@
-import { substrate } from '../../substrate/subscribe';
 import { insertAccountFollower } from '../../postgres/insert-follower';
 import { insertActivityForAccount } from '../../postgres/insert-activity';
 import { insertNotificationForOwner } from '../../postgres/notifications';
 import { SubstrateEvent, EventHandlerFn } from '../../substrate/types';
+import { substrate } from '../../connections/connect-subsocial';
 
 export const onAccountFollowed: EventHandlerFn = async (eventAction: SubstrateEvent) => {
   const { data } = eventAction;
